Move Searcher to separate feature
* Searcher as a separate feature so that it doesn't import searcher if we only need DARTS
* Copy the necessary enum from pattern API so that we don't require nightly compiler
* Spinoff searcher as a separate module
